refactor(@angular-devkit/build-angular): move babel build optimizer plugins from @angular/build
The Babel-based build optimizer plugins ('adjust-static-class-members', 'adjust-typescript-enums', 'elide-angular-metadata', and 'pure-toplevel-functions') are no longer used by '@angular/build', which now uses an OXC-based AST transform for optimization in esbuild.

This change relocates these plugins and their associated unit tests out of '@angular/build' and directly into '@angular-devkit/build-angular', which is their sole remaining consumer.

Moving these plugins also allows removing '@babel/helper-annotate-as-pure' and '@babel/helper-split-export-declaration' as direct dependencies of '@angular/build'.
