+++ packages/angular_devkit/build_angular/src/tools/babel/plugins/adjust-static-class-members.ts
