+++ packages/angular_devkit/build_angular/src/tools/babel/plugins/adjust-typescript-enums.ts
