+++ packages/angular_devkit/build_angular/src/tools/babel/plugins/elide-angular-metadata.ts
