+++ packages/angular_devkit/build_angular/src/tools/babel/plugins/pure-toplevel-functions_spec.ts
